requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
azure-functions==1.17.0
backoff==2.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cmake-format==0.6.13
cmakelang==0.6.13
cohere==4.39
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.8.0
duckdb==0.8.1
exceptiongroup==1.2.0
executing==2.0.1
Faker==19.13.0
fastapi==0.105.0
fastavro==1.9.2
fiona==1.9.5
flasgger==0.9.7.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
fonttools==4.47.0
frozenlist==1.4.1
geopandas==0.14.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kaleido==0.2.1
kiwisolver==1.4.5
lida==0.0.10
limits==3.7.0
llmx==0.0.18a0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mdurl==0.1.2
mistune==3.0.2
mizani==0.9.3
multidict==6.0.4
mysql-connector-python==8.2.0
networkx==3.2.1
numpy==1.26.2
openai==0.28.0
ordered-set==4.1.0
packaging==23.2
pandas==1.5.3
pandasai==1.5.11
parso==0.8.3
patsy==0.5.4
pexpect==4.9.0
Pillow==10.1.0
plotly==5.18.0
plotnine==0.12.4
prompt-toolkit==3.0.43
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
pydantic_core==2.14.5
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2022.10.31
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
scipy==1.11.4
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.1
tenacity==8.2.3
tiktoken==0.5.2
toolz==0.12.0
tqdm==4.66.1
traitlets==5.14.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
wcwidth==0.2.12
Werkzeug==3.0.1
wordcloud==1.9.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
